requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
mysqlclient==2.2.0
dj-database-url==2.1.0